extracting metadata program and some other classes
