+++ core/comment_list.h
@@ -1,0 +1,45 @@
+/***************************************************************************
+ *
+ * Authors:    David Strelak (davidstrelak@gmail.com)
+ *
+ * Unidad de  Bioinformatica of Centro Nacional de Biotecnologia , CSIC
+ *
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation; either version 2 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program; if not, write to the Free Software
+ * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA
+ * 02111-1307  USA
+ *
+ *  All comments concerning this program package may be sent to the
+ *  e-mail address 'xmipp@cnb.csic.es'
+ ***************************************************************************/
+
+#ifndef CORE_COMMENT_LIST_H_
+#define CORE_COMMENT_LIST_H_
+
+#include "xmipp_strings.h"
+
+/** Just a class for holding comments */
+class CommentList
+{
+public:
+    StringVector comments;
+    std::vector<int> visibility;
+    std::vector<bool> wikiVerbatim;
+
+    void addComment(const String &comment, int visible = 0, bool wikiVerbatim=false);
+    void addComment(const char * comment, bool verbatim=false);
+    void clear();
+    size_t size() const;
+};
+
+#endif /* CORE_COMMENT_LIST_H_ */
