+++ core/metadata_writemode.h
@@ -1,0 +1,38 @@
+/***************************************************************************
+ *
+ * Authors:    David Strelak (davidstrelak@gmail.com)
+ *
+ * Unidad de  Bioinformatica of Centro Nacional de Biotecnologia , CSIC
+ *
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation; either version 2 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program; if not, write to the Free Software
+ * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA
+ * 02111-1307  USA
+ *
+ *  All comments concerning this program package may be sent to the
+ *  e-mail address 'xmipp@cnb.csic.es'
+ ***************************************************************************/
+
+#ifndef CORE_METADATA_WRITEMODE_H_
+#define CORE_METADATA_WRITEMODE_H_
+
+/** Write mode
+ */
+typedef enum
+{
+    MD_OVERWRITE, //forget about the old file and overwrite it
+    MD_APPEND     //append a data_ at the file end or replace an existing one
+} WriteModeMetaData;
+
+
+#endif /* CORE_METADATA_WRITEMODE_H_ */
